+++ packages/blade/src/components/Collapsible/componentIds.ts
@@ -1,0 +1,6 @@
+export const componentIds = {
+  Collapsible: 'Collapsible',
+  CollapsibleBody: 'CollapsibleBody',
+  CollapsibleButton: 'CollapsibleButton',
+  CollapsibleLink: 'CollapsibleLink',
+};