+++ packages/blade/src/components/DatePicker/types.ts
@@ -210,6 +210,19 @@ type DatePickerInputProps = DatePickerCommonInputProps &
     placeholder?: string;
   };
 
+type DatePickerFilterChipProps = DatePickerInputProps;
+
+type FilterChipDatePickerProps = Omit<DatePickerProps<'single' | 'range'>, 'label'> & {
+  /**
+   * Sets the label for the input element.
+   */
+  label: string;
+  /**
+   *
+   * Callback which fires when clear button is clicked
+   */
+  onClearButtonClick?: () => void;
+};
 export type {
   CalendarProps,
   DatePickerProps,
@@ -218,4 +231,6 @@ export type {
   DateValue,
   DateSelectionType,
   DatePickerInputProps,
+  DatePickerFilterChipProps,
+  FilterChipDatePickerProps,
 };