+++ packages/blade/src/components/Drawer/types.ts
@@ -1,3 +1,4 @@
+import type React from 'react';
 import type { DataAnalyticsAttribute, TestID } from '~utils/types';
 
 type DrawerProps = {
@@ -53,7 +54,7 @@ type DrawerHeaderProps = {
   /**
    * Title of the Drawer
    */
-  title: string;
+  title?: string;
 
   /**
    * Subtitle of the Drawer
@@ -80,6 +81,11 @@ type DrawerHeaderProps = {
    * Link, Button[]
    */
   trailing?: React.ReactNode;
+
+  /**
+   * Children elements to be rendered inside the header
+   */
+  children?: React.ReactElement | React.ReactElement[];
 } & DataAnalyticsAttribute;
 
 export type { DrawerProps, DrawerHeaderProps };