+++ packages/blade/src/components/Dropdown/useDropdown.ts
@@ -79,9 +79,12 @@ type DropdownContextType = {
     | 'DropdownIconButton'
     | 'AutoComplete'
     | 'DropdownLink'
-    | 'SearchInput';
+    | 'SearchInput'
+    | 'FilterChipSelectInput';
+
   /** ref of triggerer. Used to call focus in certain places */
   triggererRef: React.MutableRefObject<HTMLButtonElement | null>;
+  headerAutoCompleteRef: React.MutableRefObject<HTMLButtonElement | null>;
   triggererWrapperRef: React.MutableRefObject<ContainerElementType | null>;
   actionListItemRef: React.RefObject<HTMLDivElement | null>;
   isTagDismissedRef: React.RefObject<{ value: boolean } | null>;
@@ -96,10 +99,15 @@ type DropdownContextType = {
   setHasFooterAction: (value: boolean) => void;
 
   /**
-   * Apart from dropdownTriggerer prop, we also set this boolean because in BottomSheet, the initial trigger can be Select but also have autocomplete inside of it
+   * Whether the FilterChipSelectInput is uncontrolled
    */
-  hasAutoCompleteInBottomSheetHeader: boolean;
-  setHasAutoCompleteInBottomSheetHeader: (value: boolean) => void;
+  hasUnControlledFilterChipSelectInput: boolean;
+  setHasUnControlledFilterChipSelectInput: (value: boolean) => void;
+  /**
+   * Apart from dropdownTriggerer prop, we also set this boolean because in BottomSheetHeader and DropdownHeader, the trigger can be Select but they can also have autocomplete inside of header
+   */
+  hasAutoCompleteInHeader: boolean;
+  setHasAutoCompleteInHeader: (value: boolean) => void;
 
   /**
    * A value that can be used in dependency array to know when Dropdown value is changed.
@@ -141,19 +149,24 @@ const DropdownContext = React.createContext<DropdownContextType>({
   setShouldIgnoreBlurAnimation: noop,
   hasFooterAction: false,
   setHasFooterAction: noop,
-  hasAutoCompleteInBottomSheetHeader: false,
-  setHasAutoCompleteInBottomSheetHeader: noop,
+  hasAutoCompleteInHeader: false,
+  setHasAutoCompleteInHeader: noop,
   isKeydownPressed: false,
   setIsKeydownPressed: noop,
   changeCallbackTriggerer: 0,
   setChangeCallbackTriggerer: noop,
   isControlled: false,
   setIsControlled: noop,
+  hasUnControlledFilterChipSelectInput: false,
+  setHasUnControlledFilterChipSelectInput: noop,
   dropdownBaseId: '',
   actionListItemRef: {
     current: null,
   },
   triggererRef: {
+    current: null,
+  },
+  headerAutoCompleteRef: {
     current: null,
   },
   isTagDismissedRef: {
@@ -329,7 +342,7 @@ const useDropdown = (): UseDropdownReturnValue => {
     const newIndex = index ?? activeIndex;
     let updatedIndex: number;
     const hasAutoComplete =
-      rest.hasAutoCompleteInBottomSheetHeader ||
+      rest.hasAutoCompleteInHeader ||
       dropdownTriggerer === dropdownComponentIds.triggers.AutoComplete;
     if (hasAutoComplete && filteredValues.length > 0) {
       // When its autocomplete, we don't loop over all options. We only loop on filtered options
@@ -374,13 +387,20 @@ const useDropdown = (): UseDropdownReturnValue => {
     e: React.MouseEvent<HTMLButtonElement> | React.KeyboardEvent<HTMLInputElement>,
     index: number,
   ): void => {
+    setIsKeydownPressed(false);
     const actionType = getActionFromKey(e, isOpen, dropdownTriggerer);
     if (typeof actionType === 'number') {
       onOptionChange(actionType, index);
     }
     selectOption(index);
+
     if (!isReactNative()) {
-      rest.triggererRef.current?.focus();
+      if (rest.hasAutoCompleteInHeader) {
+        // move focus to autocomplete
+        rest.headerAutoCompleteRef.current?.focus();
+      } else {
+        rest.triggererRef.current?.focus();
+      }
     }
   };
 
@@ -396,7 +416,7 @@ const useDropdown = (): UseDropdownReturnValue => {
     setIsOpen(true);
 
     if (
-      rest.hasAutoCompleteInBottomSheetHeader ||
+      rest.hasAutoCompleteInHeader ||
       dropdownTriggerer === dropdownComponentIds.triggers.AutoComplete
     ) {
       return;